Transform internal routing table into HAProxy config structure
The routing table can contain invalid entries (e.g. a port = 0 or a
hostname that is not a valid DNS identifier). The current implementation
validates the routing table as it serializes it into buffer representing
the generated HAProxy config file.

But this config file has a more complicated structure when it supports
SNI routes: servers have to be grouped together in a named `backend`
stanza, which is referred to in a separate `frontend` stanza.

This commit introduces a model that more closely resembles the structure
of the HAProxy config file. It also introduces a method that creates
this structure from a RoutingTable struct, pruning out & logging bad
entries as it goes.

This new structure simplifies the subsequent commit, which will consume
this validated, transformed model and emit a string with the HAProxy
configuration.